changed GET to POST for compiler, and put compiler url in env
