+++ .env.sample
@@ -11,3 +11,6 @@ SV_PATH_FILES=/var/shared/files
 # Ports
 BACKEND_HOST=3000
 COMPILER_HOST=3001
+
+# URLS
+COMPILER_URL=http://compiler-ms:3001
